refactor(qrscanner): broke down the qrScanner into a container and component
re #1485
